+++ lib/src/relationships/one_to_many.dart
@@ -1,6 +1,18 @@
 import 'package:wazeloquent/wazeloquent.dart';
 
 mixin OneToMany on Model {
+  /// If `foreignKey` is not specified, `foreignKey` will be issumed as the first column which contains `parent` table name the in the child table.
+  /// e.g for a car belonging to a user, foreignKey will be assumed as `user_id` which includes the name of parentTable.
+  ///
+  /// If `parentKey` is not specified, `parentKey` will be assumed as the word which is after `_`(underscore) of `foreginKey`. For example, in the case of foreign key `user_id`, `id` will be `parentKey`.
+  ///
+  /// ```dart
+  /// class Car extends Model with OneToOne{
+  ///  Future<Car?> getUser() async {
+  ///     var user = await belongsTo('users');
+  ///   }
+  /// }
+  /// ```
   Future<Map<String, Object?>?> belongsTo(String table,
       {String? foreignKey, String? parentKey}) async {
     String? q;
@@ -47,10 +59,22 @@ mixin OneToMany on Model {
       }
       return null;
     } catch (e) {
-      throw Exception(e.toString() + '\n $q');
+      throw Exception('Generated query: "$q" \n' + e.toString());
     }
   }
 
+  /// If `foreignKey` is not specified, `foreignKey` will be issumed as the first column which contains `parent` table name the in the child table.
+  /// e.g for a user having **a car or many cars**, foreignKey will be assumed as `user_id` which includes the name of parentTable in child table.
+  ///
+  /// If `parentKey` is not specified, `parentKey` will be assumed as the word which is after `_`(underscore) of `foreginKey`. For example, in the case of foreign key `user_id`, `id` will be `parentKey`.
+  ///
+  /// ```dart
+  /// class User extends Model with OneToOne{
+  ///  Future<Car?> getCar() async {
+  ///     var car = await hasMany('cars');
+  ///   }
+  /// }
+  /// ```
   Future<List<Map<String, Object?>>> hasMany(String table,
       {String? foreignKey, String? parentKey}) async {
     String? q;
@@ -95,7 +119,7 @@ mixin OneToMany on Model {
       var results = await _db.rawQuery(q);
       return results;
     } catch (e) {
-      throw Exception(e.toString() + '\n $q');
+      throw Exception('Generated query: "$q" \n' + e.toString());
     }
   }
 }